Major network video changes made the network video receiver a class made receiver able to keep track of frame quality and produce noiseless frames at lower framerate

+++ include/robosub/util.h
@@ -80,12 +80,12 @@ namespace robosub {
 	class MovingAverage{
 		int queueLen;
 		int queuePos;
-		float *queueData;
-		float curSum;
+		int *queueData;
+		int curSum;
 		
 		public:
 		MovingAverage(int);
-		void insertData(float);
+		void insertData(int);
 		void clearData();
 		float getAverage();
 	};
